ignore idea
add contacts.html
delete ap.py

add main.py
