--- ap.py
@@ -1,56 +1,0 @@
-import urllib
-from http.server import BaseHTTPRequestHandler, HTTPServer
-
-hostName = "localhost"  # Адрес для доступа по сети
-serverPort = 1111  # Порт для доступа по сети
-
-
-class MyServer(BaseHTTPRequestHandler):
-    """
-    Специальный класс, который отвечает за
-    обработку входящих запросов от клиентов
-    """
-
-    def do_GET(self):
-        """Метод для обработки входящих GET-запросов"""
-        self.send_response(200)  # Отправка кода ответа
-        self.send_header("Content-type", "text/html")  # Отправка типа данных, который будет передаваться
-        self.end_headers()  # Завершение формирования заголовков ответа
-        with open("contacts.html", "r", encoding="utf-8") as file:
-            page = file.read()
-        self.wfile.write(bytes(page, "utf-8"))  # Тело ответа
-
-    def do_POST(self):
-        content_length = int(self.headers["Content-Length"])
-        post_data = self.rfile.read(content_length)
-        parsed_data = urllib.parse.parse_qs(post_data.decode("utf-8"))
-
-        name = parsed_data.get("name", [None])[0]
-        email = parsed_data.get("email", [None])[0]
-        message = parsed_data.get("message", [None])[0]
-
-        print(f"Имя: {name}")
-        print(f"Почта: {email}")
-        print(f"Сообщение: {message}")
-
-        # Отправляем ответ клиенту
-        self.send_response(200)
-        self.send_header("Content-type", "text/html; charset=utf-8")
-        self.end_headers()
-        self.wfile.write("Форма успешно отправлена!".encode("utf-8"))
-
-
-if __name__ == "__main__":
-    webServer = HTTPServer((hostName, serverPort), MyServer)
-    print("Server started http://%s:%s" % (hostName, serverPort))
-
-    try:
-        # Cтарт веб-сервера в бесконечном цикле прослушивания входящих запросов
-        webServer.serve_forever()
-    except KeyboardInterrupt:
-        # Корректный способ остановить сервер в консоли через сочетание клавиш Ctrl + C
-        pass
-
-    # Корректная остановка веб-сервера, чтобы он освободил адрес и порт в сети, которые занимал
-    webServer.server_close()
-    print("Server stopped.")
